refactor(cli): extract standards cluster into commands/standards.ts (5.0-alpha)
Seventh and final cluster of the 5.0-alpha cli.ts god-module split.

- New src/commands/standards.ts: cmdStandards + cmdBaseline (exported),
  cmdStandardsFreeze + the freezeStandardsBaseline helper (module-private;
  freezeStandardsBaseline is shared by `standards freeze` and `baseline freeze`,
  both now in this module).
- cmdReview stays in cli.ts — it orchestrates cmdCheck + cmdDesign (still inline);
  it calls the now-imported cmdStandards. Deferred to a later PR with setup Phase B.
- cli.ts imports cmdStandards + cmdBaseline back; COMMANDS table unchanged; prunes
  the now-unused Baseline type import.
- Dynamic imports re-pathed ./ -> ../.

cli.ts: 3789 -> 3613 lines (-176; -2717 cumulative across PRs 1-7, from the
original 6330 — ~43% smaller). Typecheck, eslint (0 errors), full suite (2580),
prettier all green; kit standards / baseline / review smoke-tested.
